Changes to the file

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,8 +8,8 @@ size_t print_listint_safe(const listint_t *head)
 {
 	const listint_t *cur = head;
 	const listint_t *tmp = head;
-	unsigned int elements = 0;
-	unsigned int tmp_index = 0;
+	size_t elements = 0;
+	size_t tmp_index = 0;
 
 	while (cur)
 	{
